requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.5.0
pydantic-settings>=2.2.1
python-multipart>=0.0.9
requests>=2.31.0
httpx>=0.27.0
pytest>=7.4.0
pytest-anyio>=0.0.0
ruff>=0.4.0
black>=24.0.0
chromadb>=0.5.0
sentence-transformers>=2.2.2
numpy>=1.26.0
scikit-learn>=1.3.0
neo4j>=5.14.0
pdfminer.six>=20220524
python-dotenv>=1.0.0
